Finish to explode Gruntfile
