+++ grunt_tasks/config/webfont.js
@@ -1,4 +1,4 @@
-module.exports = function(grunt, happyplan) {
+module.exports = function(grunt) {
   "use strict";
 
   return {
@@ -8,10 +8,8 @@ module.exports = function(grunt, happyplan) {
       destCss: '<%= happyplan.theme.local.assets.styles %>',
       options: {
           relativeFontPath: require('path').relative(
-            // we must process template here because it's not already done by the grunt.init at this time
-            // PR if u have a better solution :)
-            __dirname + '/' + grunt.template.process('<%= happyplan.dist.assets.styles %>', { data: { happyplan: happyplan}}),
-            __dirname + '/' + grunt.template.process('<%= happyplan.dist.assets.fonts %>', { data: { happyplan: happyplan}})
+            __dirname + '/' + grunt.config.get(['happyplan', 'dist', 'assets', 'styles']),
+            __dirname + '/' + grunt.config.get(['happyplan', 'dist', 'assets', 'fonts'])
           ),
           stylesheet: 'scss',
           hashes: false,
