+++ grunt_tasks/newpost.js
@@ -1,13 +1,21 @@
-module.exports = function(grunt, happyplan) {
+module.exports = function(grunt) {
   "use strict";
+
+  function unquote(str) {
+    if (str.indexOf('"') === 0 && str.lastIndexOf('"') === str.length-1 ||
+        str.indexOf("'") === 0 && str.lastIndexOf("'") === str.length-1) {
+      return str.slice(1,str.length-1);
+    }
+    return str;
+  }
 
   grunt.registerTask('happyplan:newpost', 'Generate a new file ready for a fresh post', function () {
     var postName = grunt.option('name') || 'change-me';
-    postName = grunt.util._.unquote(postName)
+    postName = unquote(postName)
 
     // support tag & tags option
     var tags = grunt.option('tag') || grunt.option('tags') || "";
-    tags = grunt.util._.unquote(tags).split(',')
+    tags = unquote(tags).split(',')
 
     // Template
     var fileContent = [
