+++ test/features/theme_child/expected/test/index.html
@@ -1,1 +1,2 @@
+
 <p>This is a test</p>
