+++ test/features/theme_parent/expected/2013/test.html
@@ -1,1 +1,2 @@
+
 <p>This is a test</p>
